refactor: 무한 스크롤 옵저버 훅 분리
